requirements.txt
fastapi
uvicorn
python-multipart
google-genai
websockets
httpx
